"Replaced" walls with tiles and resized room
Room size is changed to fit on 1080p monitor
Made canvas lock on the screen
Walls are now derived from the tiles in a room. Tiles are derived from the walls that are originally generated. These walls are offset by roomoffset towards the center of the room so the walls go through the center of the tiles, not the edges.
room bitmap is shifted to make it so you can walk into walls a bit